Move the edit action into the site editor
